requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0

# Database & ORM
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
alembic==1.12.1

# Caching & Performance
redis==5.0.1
hiredis==2.2.3
Flask-Caching==2.1.0

# Security & Auth
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
bcrypt==4.1.1

# CORS & Rate Limiting
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Response Compression
Flask-Compress==1.14

# Face Recognition & ML
deepface==0.0.79
opencv-python-headless==4.8.1.78
tensorflow==2.15.0
tf-keras==2.15.0
numpy==1.24.3
Pillow==10.1.0
mtcnn==0.1.1

# Vector Search
faiss-cpu==1.7.4

# Background Tasks (Optional)
celery==5.3.4
flower==2.0.1

# Email (Optional)
Flask-Mail==0.9.1

# Monitoring & Logging
sentry-sdk[flask]==1.38.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Image Processing
scikit-image==0.22.0

# Validation
marshmallow==3.20.1
webargs==8.3.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==20.0.3

# Development
ipython==8.18.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1